refactor: homeSection을 컨테이너 디렉토리로 위치 수정
- MasonryImageList 컴포넌트를 Section으로 합치도록 수정함
